Refactor QueryTermManager into abstract class and implement SimpleQueryTermManager
