+++ Assets/Scenes/SimpleInputScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8cb51fa08b0f54d588be86efb38db64d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
