+++ Assets/Scripts/VitrivrVR/Query/Term/SimpleQueryTermManager.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 959fa48eef01484ba9b7af7bd94199ae
+timeCreated: 1716300145
